Add files via upload
This is an algorithm used to automatically extract CT dose screen information from SIEMENs CT scanner model using Optical character recognition Method. This algorithm works perfectly and able to analyze almost 2000 CT dose screen in an hour. The CT scanner model used to train this algorithm was the SOMATOM Definition and the SOMATOM Definition AS+. 

Python packages needed to run this algorithm :
1) Pydicom
2) Threading 
3) OpenCV
4) Numpy
5) MatPlotLib
6) TesseractOCR - download here (https://github.com/tesseract-ocr/tesseract) 
7) Pytesseract
8) Regex